Migrate balance-replicas to 'api' module
