+++ solr/core/src/java/org/apache/solr/handler/admin/api/BalanceReplicas.java
@@ -16,7 +16,6 @@
  */
 package org.apache.solr.handler.admin.api;
 
-import static org.apache.solr.client.solrj.impl.BinaryResponseParser.BINARY_CONTENT_TYPE_V2;
 import static org.apache.solr.cloud.Overseer.QUEUE_OPERATION;
 import static org.apache.solr.common.params.CollectionParams.NODES;
 import static org.apache.solr.common.params.CommonAdminParams.ASYNC;
@@ -24,47 +23,35 @@ import static org.apache.solr.common.params.CommonAdminParams.WAIT_FOR_FINAL_STA
 import static org.apache.solr.handler.admin.CollectionsHandler.DEFAULT_COLLECTION_OP_TIMEOUT;
 import static org.apache.solr.security.PermissionNameProvider.Name.COLL_EDIT_PERM;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
 import javax.inject.Inject;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
+import org.apache.solr.client.api.endpoint.BalanceReplicasApi;
+import org.apache.solr.client.api.model.BalanceReplicasRequestBody;
 import org.apache.solr.client.api.model.SolrJerseyResponse;
 import org.apache.solr.client.solrj.SolrResponse;
 import org.apache.solr.common.cloud.ZkNodeProps;
 import org.apache.solr.common.params.CollectionParams.CollectionAction;
 import org.apache.solr.core.CoreContainer;
 import org.apache.solr.handler.admin.CollectionsHandler;
-import org.apache.solr.jersey.JacksonReflectMapWriter;
 import org.apache.solr.jersey.PermissionName;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
 
 /** V2 API for balancing the replicas that already exist across a set of nodes. */
-@Path("cluster/replicas/balance")
-public class BalanceReplicasAPI extends AdminAPIBase {
+public class BalanceReplicas extends AdminAPIBase implements BalanceReplicasApi {
 
   @Inject
-  public BalanceReplicasAPI(
+  public BalanceReplicas(
       CoreContainer coreContainer,
       SolrQueryRequest solrQueryRequest,
       SolrQueryResponse solrQueryResponse) {
     super(coreContainer, solrQueryRequest, solrQueryResponse);
   }
 
-  @POST
-  @Produces({"application/json", "application/xml", BINARY_CONTENT_TYPE_V2})
+  @Override
   @PermissionName(COLL_EDIT_PERM)
-  @Operation(summary = "Balance Replicas across the given set of Nodes.")
-  public SolrJerseyResponse balanceReplicas(
-      @RequestBody(description = "Contains user provided parameters")
-          BalanceReplicasRequestBody requestBody)
+  public SolrJerseyResponse balanceReplicas(BalanceReplicasRequestBody requestBody)
       throws Exception {
     final SolrJerseyResponse response = instantiateJerseyResponse(SolrJerseyResponse.class);
     final CoreContainer coreContainer = fetchAndValidateZooKeeperAwareCoreContainer();
@@ -95,34 +82,5 @@ public class BalanceReplicasAPI extends AdminAPIBase {
     remoteMessage.put(QUEUE_OPERATION, CollectionAction.BALANCE_REPLICAS.toLower());
 
     return new ZkNodeProps(remoteMessage);
-  }
-
-  public static class BalanceReplicasRequestBody implements JacksonReflectMapWriter {
-
-    public BalanceReplicasRequestBody() {}
-
-    public BalanceReplicasRequestBody(Set<String> nodes, Boolean waitForFinalState, String async) {
-      this.nodes = nodes;
-      this.waitForFinalState = waitForFinalState;
-      this.async = async;
-    }
-
-    @Schema(
-        description =
-            "The set of nodes across which replicas will be balanced. Defaults to all live data nodes.")
-    @JsonProperty(value = "nodes")
-    public Set<String> nodes;
-
-    @Schema(
-        description =
-            "If true, the request will complete only when all affected replicas become active. "
-                + "If false, the API will return the status of the single action, which may be "
-                + "before the new replica is online and active.")
-    @JsonProperty("waitForFinalState")
-    public Boolean waitForFinalState = false;
-
-    @Schema(description = "Request ID to track this action which will be processed asynchronously.")
-    @JsonProperty("async")
-    public String async;
   }
 }
